fix cpu profiler not saving file when stopped
